+++ server/server.ts
@@ -3,11 +3,11 @@ import socketIO = require("socket.io");
 
 import { Spell } from "../common/Spell";
 import { StatBlock } from "../common/StatBlock";
+import { InMemoryPlayerViewManager } from "./InMemoryPlayerViewManager";
 import * as DB from "./dbconnection";
 import { getDbConnectionString } from "./getDbConnectionString";
 import LaunchServer from "./launchserver";
 import * as L from "./library";
-import { InMemoryPlayerViewManager } from "./playerviewmanager";
 import ConfigureRoutes from "./routes";
 import GetSessionMiddleware from "./session";
 import ConfigureSockets from "./sockets";
